refactor: `run_utility` を統合して廃止
